backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
boto3==1.29.7
openai>=1.56.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyPDF2==3.0.1
python-magic==0.4.27
pillow==10.1.0
pytesseract==0.3.10
aiofiles==23.2.1
httpx<0.28.0
